Lift util to own file
