+++ calyx-py/test/correctness/builder_util.py
@@ -1,0 +1,94 @@
+import calyx.builder as cb
+
+
+def insert_eq(comp: cb.ComponentBuilder, a, b, cell, width):
+    """Inserts wiring into component {comp} to check if {a} == {b}.
+    1. Within {comp}, creates a combinational group called {cell}_group.
+    2. Within the group, creates a {cell} that checks equalities of {width}.
+    3. Puts the values {a} and {b} into {cell}.
+    4. Returns the equality-checking cell and the overall group.
+    """
+    eq_cell = comp.eq(cell, width)
+    with comp.comb_group(f"{cell}_group") as eq_group:
+        eq_cell.left = a
+        eq_cell.right = b
+    return eq_cell, eq_group
+
+
+def insert_incr(comp: cb.ComponentBuilder, reg, cell, group):
+    """Inserts wiring into component {comp} to increment {reg} by 1.
+    1. Within component {comp}, creates a group called {group}.
+    2. Within {group}, adds a cell {cell} that computes sums.
+    3. Puts the values of {port} and 1 into {cell}.
+    4. Then puts the answer of the computation back into {port}.
+    4. Returns the group that does this.
+    """
+    incr_cell = comp.add(cell, 32)
+    with comp.group(group) as incr_group:
+        incr_cell.left = reg.out
+        incr_cell.right = 1
+        reg.write_en = 1
+        reg.in_ = incr_cell.out
+        incr_group.done = reg.done
+    return incr_group
+
+
+def insert_decr(comp: cb.ComponentBuilder, reg, cell, group):
+    """Inserts wiring into component {comp} to decrement {reg} by 1.
+    1. Within component {comp}, creates a group called {group}.
+    2. Within {group}, adds a cell {cell} that computes differences.
+    3. Puts the values of {port} and 1 into {cell}.
+    4. Then puts the answer of the computation back into {port}.
+    4. Returns the group that does this.
+    """
+    decr_cell = comp.sub(cell, 32)
+    with comp.group(group) as decr_group:
+        decr_cell.left = reg.out
+        decr_cell.right = cb.const(32, 1)
+        reg.write_en = 1
+        reg.in_ = decr_cell.out
+        decr_group.done = reg.done
+    return decr_group
+
+
+def mem_load(comp: cb.ComponentBuilder, mem, i, reg, group):
+    """Loads a value from one memory into a register.
+    1. Within component {comp}, creates a group called {group}.
+    2. Within {group}, reads from memory {mem} at address {i}.
+    3. Writes the value into register {reg}.
+    4. Returns the group that does this.
+    """
+    with comp.group(group) as load_grp:
+        mem.addr0 = i
+        reg.write_en = 1
+        reg.in_ = mem.read_data
+        load_grp.done = reg.done
+    return load_grp
+
+
+def mem_store(comp: cb.ComponentBuilder, mem, i, val, group):
+    """Stores a value from one memory into another.
+    1. Within component {comp}, creates a group called {group}.
+    2. Within {group}, reads from {val}.
+    3. Writes the value into memory {mem} at address i.
+    4. Returns the group that does this.
+    """
+    with comp.group(group) as store_grp:
+        mem.addr0 = i
+        mem.write_en = 1
+        mem.write_data = val
+        store_grp.done = mem.done
+    return store_grp
+
+
+def reg_store(comp: cb.ComponentBuilder, reg, val, group):
+    """Stores a value in a register.
+    1. Within component {comp}, creates a group called {group}.
+    2. Within {group}, sets the register {reg} to {val}.
+    3. Returns the group that does this.
+    """
+    with comp.group(group) as reg_grp:
+        reg.in_ = val
+        reg.write_en = 1
+        reg_grp.done = reg.done
+    return reg_grp
