+++ calyx-py/test/correctness/fifo.py
@@ -1,98 +1,6 @@
 # pylint: disable=import-error
+import builder_util as util
 import calyx.builder as cb
-
-
-def insert_eq(comp: cb.ComponentBuilder, a, b, cell, width):
-    """Inserts wiring into component {comp} to check if {a} == {b}.
-    1. Within {comp}, creates a combinational group called {cell}_group.
-    2. Within the group, creates a {cell} that checks equalities of {width}.
-    3. Puts the values {a} and {b} into {cell}.
-    4. Returns the equality-checking cell and the overall group.
-    """
-    eq_cell = comp.eq(cell, width)
-    with comp.comb_group(f"{cell}_group") as eq_group:
-        eq_cell.left = a
-        eq_cell.right = b
-    return eq_cell, eq_group
-
-
-def insert_incr(comp: cb.ComponentBuilder, reg, cell, group):
-    """Inserts wiring into component {comp} to increment {reg} by 1.
-    1. Within component {comp}, creates a group called {group}.
-    2. Within {group}, adds a cell {cell} that computes sums.
-    3. Puts the values of {port} and 1 into {cell}.
-    4. Then puts the answer of the computation back into {port}.
-    4. Returns the group that does this.
-    """
-    incr_cell = comp.add(cell, 32)
-    with comp.group(group) as incr_group:
-        incr_cell.left = reg.out
-        incr_cell.right = 1
-        reg.write_en = 1
-        reg.in_ = incr_cell.out
-        incr_group.done = reg.done
-    return incr_group
-
-
-def insert_decr(comp: cb.ComponentBuilder, reg, cell, group):
-    """Inserts wiring into component {comp} to decrement {reg} by 1.
-    1. Within component {comp}, creates a group called {group}.
-    2. Within {group}, adds a cell {cell} that computes differences.
-    3. Puts the values of {port} and 1 into {cell}.
-    4. Then puts the answer of the computation back into {port}.
-    4. Returns the group that does this.
-    """
-    decr_cell = comp.sub(cell, 32)
-    with comp.group(group) as decr_group:
-        decr_cell.left = reg.out
-        decr_cell.right = cb.const(32, 1)
-        reg.write_en = 1
-        reg.in_ = decr_cell.out
-        decr_group.done = reg.done
-    return decr_group
-
-
-def mem_load(comp: cb.ComponentBuilder, mem, i, reg, group):
-    """Loads a value from one memory into a register.
-    1. Within component {comp}, creates a group called {group}.
-    2. Within {group}, reads from memory {mem} at address {i}.
-    3. Writes the value into register {reg}.
-    4. Returns the group that does this.
-    """
-    with comp.group(group) as load_grp:
-        mem.addr0 = i
-        reg.write_en = 1
-        reg.in_ = mem.read_data
-        load_grp.done = reg.done
-    return load_grp
-
-
-def mem_store(comp: cb.ComponentBuilder, mem, i, val, group):
-    """Stores a value from one memory into another.
-    1. Within component {comp}, creates a group called {group}.
-    2. Within {group}, reads from {val}.
-    3. Writes the value into memory {mem} at address i.
-    4. Returns the group that does this.
-    """
-    with comp.group(group) as store_grp:
-        mem.addr0 = i
-        mem.write_en = 1
-        mem.write_data = val
-        store_grp.done = mem.done
-    return store_grp
-
-
-def reg_store(comp: cb.ComponentBuilder, reg, val, group):
-    """Stores a value in a register.
-    1. Within component {comp}, creates a group called {group}.
-    2. Within {group}, sets the register {reg} to {val}.
-    3. Returns the group that does this.
-    """
-    with comp.group(group) as reg_grp:
-        reg.in_ = val
-        reg.write_en = 1
-        reg_grp.done = reg.done
-    return reg_grp
 
 
 def insert_raise_err_if_i_eq_15(prog):
@@ -108,8 +16,8 @@ def insert_raise_err_if_i_eq_15(prog):
     i = raise_err_if_i_eq_15.input("i", 32)
     err = raise_err_if_i_eq_15.reg("err", 1, is_ref=True)
 
-    i_eq_15 = insert_eq(raise_err_if_i_eq_15, i, 15, "i_eq_15", 32)
-    raise_err = reg_store(raise_err_if_i_eq_15, err, 1, "raise_err")
+    i_eq_15 = util.insert_eq(raise_err_if_i_eq_15, i, 15, "i_eq_15", 32)
+    raise_err = util.reg_store(raise_err_if_i_eq_15, err, 1, "raise_err")
 
     raise_err_if_i_eq_15.control += [
         cb.if_(
@@ -158,29 +66,31 @@ def insert_fifo(prog):
     len = fifo.reg("len", 32, is_ref=True)  # The length of the queue
 
     # Cells and groups to compute equality
-    pop_eq_push = insert_eq(fifo, pop, push, "pop_eq_push", 1)  # `pop` == `push`
-    pop_eq_1 = insert_eq(fifo, pop, 1, "pop_eq_1", 1)  # `pop` == 1
-    push_eq_1 = insert_eq(fifo, push, 1, "push_eq_1", 1)  # `push` == 1
-    write_eq_10 = insert_eq(fifo, write.out, 10, "write_eq_10", 32)  # `write` == 10
-    read_eq_10 = insert_eq(fifo, read.out, 10, "read_eq_10", 32)  # `read` == 10
-    len_eq_0 = insert_eq(fifo, len.out, 0, "len_eq_0", 32)  # `len` == 0
-    len_eq_10 = insert_eq(fifo, len.out, 10, "len_eq_10", 32)  # `len` == 10
+    pop_eq_push = util.insert_eq(fifo, pop, push, "pop_eq_push", 1)  # `pop` == `push`
+    pop_eq_1 = util.insert_eq(fifo, pop, 1, "pop_eq_1", 1)  # `pop` == 1
+    push_eq_1 = util.insert_eq(fifo, push, 1, "push_eq_1", 1)  # `push` == 1
+    write_eq_10 = util.insert_eq(
+        fifo, write.out, 10, "write_eq_10", 32
+    )  # `write` == 10
+    read_eq_10 = util.insert_eq(fifo, read.out, 10, "read_eq_10", 32)  # `read` == 10
+    len_eq_0 = util.insert_eq(fifo, len.out, 0, "len_eq_0", 32)  # `len` == 0
+    len_eq_10 = util.insert_eq(fifo, len.out, 10, "len_eq_10", 32)  # `len` == 10
 
     # Cells and groups to increment read and write registers
-    write_incr = insert_incr(fifo, write, "add1", "write_incr")  # write++
-    read_incr = insert_incr(fifo, read, "add2", "read_incr")  # read++
-    len_incr = insert_incr(fifo, len, "add5", "len_incr")  # len++
-    len_decr = insert_decr(fifo, len, "add6", "len_decr")  # len--
+    write_incr = util.insert_incr(fifo, write, "add1", "write_incr")  # write++
+    read_incr = util.insert_incr(fifo, read, "add2", "read_incr")  # read++
+    len_incr = util.insert_incr(fifo, len, "add5", "len_incr")  # len++
+    len_decr = util.insert_decr(fifo, len, "add6", "len_decr")  # len--
 
     # Cells and groups to modify flags, which are registers
-    write_wrap = reg_store(fifo, write, 0, "write_wraparound")  # zero out `write`
-    read_wrap = reg_store(fifo, read, 0, "read_wraparound")  # zero out `read`
-    raise_err = reg_store(fifo, err, 1, "raise_err")  # set `err` to 1
-    zero_out_ans = reg_store(fifo, ans, 0, "zero_out_ans")  # zero out `ans`
+    write_wrap = util.reg_store(fifo, write, 0, "write_wraparound")  # zero out `write`
+    read_wrap = util.reg_store(fifo, read, 0, "read_wraparound")  # zero out `read`
+    raise_err = util.reg_store(fifo, err, 1, "raise_err")  # set `err` to 1
+    zero_out_ans = util.reg_store(fifo, ans, 0, "zero_out_ans")  # zero out `ans`
 
     # Load and store into an arbitary slot in memory
-    write_to_mem = mem_store(fifo, mem, write.out, payload, "write_payload_to_mem")
-    read_from_mem = mem_load(fifo, mem, read.out, ans, "read_payload_from_mem")
+    write_to_mem = util.mem_store(fifo, mem, write.out, payload, "write_payload_to_mem")
+    read_from_mem = util.mem_load(fifo, mem, read.out, ans, "read_payload_from_mem")
 
     fifo.control += [
         cb.if_(
@@ -275,14 +185,14 @@ def insert_main(prog, fifo, raise_err_if_i_eq_15):
     j = main.reg("j", 32)  # The index on the answer-list we'll write to
     command = main.reg("command", 32)  # The command we're currently processing
 
-    zero_i = reg_store(main, i, 0, "zero_i")  # zero out `i`
-    zero_j = reg_store(main, j, 0, "zero_j")  # zero out `j`
-    incr_i = insert_incr(main, i, "add3", "incr_i")  # i = i + 1
-    incr_j = insert_incr(main, j, "add4", "incr_j")  # j = j + 1
-    err_eq_zero = insert_eq(main, err.out, 0, "err_eq_0", 1)  # is `err` flag down?
-    read_command = mem_load(main, commands, i.out, command, "read_command")
-    command_eq_zero = insert_eq(main, command.out, 0, "command_eq_zero", 32)
-    write_ans = mem_store(main, ans_mem, j.out, ans.out, "write_ans")
+    zero_i = util.reg_store(main, i, 0, "zero_i")  # zero out `i`
+    zero_j = util.reg_store(main, j, 0, "zero_j")  # zero out `j`
+    incr_i = util.insert_incr(main, i, "add3", "incr_i")  # i = i + 1
+    incr_j = util.insert_incr(main, j, "add4", "incr_j")  # j = j + 1
+    err_eq_zero = util.insert_eq(main, err.out, 0, "err_eq_0", 1)  # is `err` flag down?
+    read_command = util.mem_load(main, commands, i.out, command, "read_command")
+    command_eq_zero = util.insert_eq(main, command.out, 0, "command_eq_zero", 32)
+    write_ans = util.mem_store(main, ans_mem, j.out, ans.out, "write_ans")
 
     main.control += [
         zero_i,
